box_training_methods train --task graph_modeling --data_path ./data/graphs/balanced_tree/branching=2-log_num_nodes=12-transitive_closure=False/ --model_type hard_box --dim 8 --epochs 10
